Game info and dropdown menus

+++ frontend/src/utils/date.utils.ts
@@ -1,7 +1,7 @@
 export default function formatDateTime(date: Date): string {
-  const month = date.toLocaleString('en-US', { month: 'long', timeZone: 'UTC' });
+  const month = date.toLocaleString('en-US', { month: 'short', timeZone: 'UTC' });
   const day = date.getUTCDate();
-  const weekday = date.toLocaleString('en-US', { weekday: 'long', timeZone: 'UTC' });
+  const weekday = date.toLocaleString('en-US', { weekday: 'short', timeZone: 'UTC' });
   const hours = date.getUTCHours();
   const minutes = date.getUTCMinutes();
   const ampm = hours >= 12 ? 'PM' : 'AM';
